Переносит код кнопок в модалке в отдельный файл
