+++ src/js/all-dev-parts.js
@@ -10,7 +10,19 @@ import './dev-5/to-the-top.js';
 
 import ApiService from './dev-1/api.js';
 import { renderTrends, renderLibrary, getTrailerKey } from './dev-1/renders.js';
-import { renderModal, remove, save, test, get } from './dev-3/dev-3-main.js';
+import { renderModal } from './dev-3/dev-3-main.js';
+import {
+  removeQueueId,
+  removeWatcheId,
+  addQueueIdFilm,
+  addWatcheIdFilm,
+  textModalBtn,
+  inArrayKey,
+  test,
+  remove,
+  save,
+  get,
+} from './dev-4/dev-4-main.js';
 import { renderSearch, idToGenre, genreData } from './dev-2/dev-2-main.js';
 import { renderMovies, clearFilmsGallery, moviesLibraryMarkup } from './dev-5/dev-5-main.js';
